# client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "English Reader 客户端接口",
    "version": "0.1.0",
    "description": "只含 /v1/client。由 scripts/export_contract.py 导出，不要手改。\n\n**只增不减**（架构铁律 5）：可以加字段，不能删字段或改字段含义。侧载安装的客户端更新滞后，手机上大概率长期跑着旧版本。\n\n**capabilities 区分「没数据」和「没实现」**：一个 null 字段本身是歧义的，`root_known: null` 可能是「这个词没有词根」，也可能是「水平估计还没做」，而客户端要把这两种画成不同的样子。\n\n**幂等键重复上报算正常，不算错误**：离线客户端会把拿不准的都重发一遍。\n\n**今日包不含真题**：真题走 /library?source=cet4|cet6|kaoyan。"
  },
  "paths": {
...
    "/v1/client/sentences": {
      "get": {
        "tags": [
          "client:review"
        ],
        "summary": "在学的那些词的句子（不分池）",
        "description": "你在学的每个词，连它的全部句子——**服务端不分池**。\n\n**P9 §11:那条线把两件事分开了。** 造句子要钱、要模型、要 24 小时醒着，\n是工厂的活；而「这一句该当考题还是当提示」取决于你读完过哪些文章、\n见过哪些句子——那是学习记录，现在在设备上。所以这里原样全给，\n由客户端分（`ERCore/SentencePool`，三条规则逐条镜像 `sentences.split_pools`）。\n\n**依据是你上报的词池快照**（§7），不是服务端自己推的。服务端对学习记录只有\n两种关系:生文需要的那一小撮信号，和它不解释的存档——这里用的是前者，\n而它连解释都不算:直接用。**所以还没报过快照的设备会拿到空列表**，\n而响应里的 `reported_at` 为空正是在说这件事:不是「你没在学任何词」，\n是「服务端还不知道」。\n\n**只给词，不给词组。** 复习有意跳过词组（`verify_phase3` 2.2），\n而句子池本来也是按词与义项建的——词组拿不到句子。\n\n**它取代了 `/reviews`，而不是补充它。** 那个端点带着队列、方向、权重、进度，\n全是学习状态；它在同一个 Phase 删掉了（§11）。这一个只带内容。",
        "operationId": "sentence_pool_v1_client_sentences_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentencePoolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/client/reviews/spelling": {
      "post": {
        "tags": [
          "client:review"
        ],
        "summary": "上报一次拼写",
        "description": "Recorded, never scheduled on.\n\n决定 13 keeps spelling out of the scheduler entirely: it is an optional\nreinforcement, and letting it move due dates would mix a productive-recall\ndifficulty into a recognition track. What is stored is what was typed, not\njust whether it was right — that is the whole value of the record.",
        "operationId": "report_spelling_v1_client_reviews_spelling_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/client/me": {
      "get": {
        "tags": [
          "client:today"
        ],
        "summary": "这把令牌是谁的（测试连接用）",
        "operationId": "me_v1_client_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Capabilities": {
        "properties": {
          "level_estimate": {
            "type": "boolean",
            "title": "Level Estimate",
            "description": "水平估计。它填的是 learner.level、difficulty_for_you 和 root_known"
          },
          "memory_state": {
            "type": "boolean",
            "title": "Memory State",
            "description": "每个义项的记忆状态与到期时间"
          },
          "proper_noun_notes": {
            "type": "boolean",
            "title": "Proper Noun Notes",
            "description": "专有名词的一句话说明"
          },
          "exam_frequency": {
            "type": "boolean",
            "title": "Exam Frequency",
            "description": "真题考频。为假时 exam 整块缺席而不是为零——零会被读成「真题里从没出现过」，那是另一回事"
          },
          "phrases": {
            "type": "boolean",
            "title": "Phrases",
            "description": "词组识别跑过没有。为假时，空的 phrases 列表意思是「还没看过」而不是「这篇没有词组」"
          }
        },
        "type": "object",
        "required": [
          "level_estimate",
          "memory_state",
          "proper_noun_notes",
          "exam_frequency",
          "phrases"
        ],
        "title": "Capabilities",
        "description": "Which reserved slots actually carry values yet.\n\nA null alone is ambiguous: `root_known: null` could mean \"this word has no\nroot\" or \"the ability estimate does not exist\", and a client renders those\ntwo differently — nothing at all, versus 功能待开发. The distinction belongs\nin the contract rather than being guessed from the value."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Learner": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "词汇量估计。水平评估（主文档 §A）尚未实现，所以现在恒为 null——由 capabilities.level_estimate 区分「没数据」和「没实现」。"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Learner",
        "description": "Who this response is about, echoed at the top of every client response.\n\nLets a client notice it is looking at someone else's cache and drop it. The\nlearner is derived from the device token — no learner id appears in a path."
      },
...
      "MeResponse": {
        "properties": {
          "learner": {
            "$ref": "#/components/schemas/Learner"
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          }
        },
        "type": "object",
        "required": [
          "learner",
          "capabilities"
        ],
        "title": "MeResponse",
        "description": "这把令牌是谁的，以及哪些留好的位置真的有值了。\n\n**测试连接用的就是它。** 在这之前那个探针打的是日历那个端点\n（「最轻的一个」），而日历 P9 搬到了设备上——探针因此需要一个真正最轻的:\n不读学习记录、不组装任何东西、不下发一个字节的内容。\n\n**它同时回答了探针真正关心的两件事**:令牌认不认（401 还是 200），\n以及对面是谁（换过令牌之后名字对不对）。"
      },
...
      "SenseCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "headword": {
            "type": "string",
            "title": "Headword"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "pos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos"
          },
          "concept_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept En"
          },
          "gloss_zh": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gloss Zh"
          },
          "exam_frequency": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Frequency",
            "description": "真题考频"
          }
        },
        "type": "object",
        "required": [
          "id",
          "headword",
          "ordinal"
        ],
        "title": "SenseCard",
        "description": "One sense, as a review card shows it."
      },
...
      "SentencePoolResponse": {
        "properties": {
          "learner": {
            "$ref": "#/components/schemas/Learner"
          },
          "reported_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported At",
            "description": "那份快照是什么时候的。**为空表示这台设备还没报过**——那时 `items` 也是空的，而那不是「你没在学任何词」，是「服务端还不知道」"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyItemSentences"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "learner",
          "items"
        ],
        "title": "SentencePoolResponse",
        "description": "在学的那些词的句子。\n\n**依据是你上报的词池快照**（§7），不是服务端自己推的——\n服务端对学习记录只有两种关系:生文需要的那一小撮信号，和它不解释的存档。\n这里用的是前者，而它连解释都不算:直接用。"
      },
...
      "SpellingIn": {
        "properties": {
          "item_key": {
            "type": "string",
            "title": "Item Key"
          },
          "typed": {
            "type": "string",
            "title": "Typed"
          }
        },
        "type": "object",
        "required": [
          "item_key",
          "typed"
        ],
        "title": "SpellingIn"
      },
...
      "SpellingResponse": {
        "properties": {
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "expected": {
            "type": "string",
            "title": "Expected"
          }
        },
        "type": "object",
        "required": [
          "correct",
          "expected"
        ],
        "title": "SpellingResponse",
        "description": "Recorded, never scheduled on.\n\nSpelling stays out of the scheduler entirely: it is optional reinforcement,\nand letting it move due dates would mix a productive-recall difficulty into\na recognition track. What is stored is what was *typed*, not just whether it\nwas right — that is the whole value of the record."
      },
...
      "StudyItemSentences": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "item_key": {
            "type": "string",
            "title": "Item Key"
          },
          "sense_id": {
            "type": "integer",
            "title": "Sense Id"
          },
          "word": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "sense": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SenseCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/SentenceCard"
            },
            "type": "array",
            "title": "Sentences",
            "description": "这个词（这个义项）的全部句子，**没有分池**。`source` 与 `sentence_id`／`article_id` 够客户端自己分"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "item_key",
          "sense_id",
          "sentences"
        ],
        "title": "StudyItemSentences",
        "description": "一个在学的词，连它的句子——**不分池**。\n\n**P9 §11。** 分池（哪句当考题、哪句当提示）依赖「你读完过哪些文章、\n见过哪些句子」，那是学习记录；而造句子要钱、要模型，是内容生产。\n那条线把两件事分开了，所以这里原样全给，由客户端分（`ERCore/SentencePool`）。\n\n**和 `ReviewItem` 的差别正是那条线**：这里没有 `queue_id`、`bucket`、\n`direction`、`asks`、`weight`、`done`——那些全是学习状态，现在由设备重放算出来。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WordCard": {
        "properties": {
          "headword": {
            "type": "string",
            "title": "Headword"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonetic"
          },
          "translation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation",
            "description": "词典那一堆逗号隔开的释义。义项集缺席时的退路"
          },
          "senses": {
            "items": {
              "$ref": "#/components/schemas/WordSense"
            },
            "type": "array",
            "title": "Senses"
          }
        },
        "type": "object",
        "required": [
          "headword"
        ],
        "title": "WordCard",
        "description": "The whole word, as the reveal screen shows it.\n\nDuplicates what the article glossary carries, and that is the point:\n**the review payload has to stand on its own.** Article bodies are cleared\none at a time (P5 决定 10) and the word may have been met months ago, so a\nclient reaching into a cached article for this would lose half its card\nwhenever the cache was tidied — silently."
      },
...
    }
  }
}